+++ src/hull/commands/dispatch.c
@@ -42,6 +42,7 @@
 #ifdef HL_ENABLE_HTTP_CLIENT
 #include "hull/commands/update.h"
 #include "hull/commands/tools.h"
+#include "hull/commands/platform.h"
 #endif
 #include "hull/commands/sign_release.h"
 #include "hull/commands/verify_release.h"
@@ -86,6 +87,7 @@ static const HlCommand commands[] = {
 #ifdef HL_ENABLE_HTTP_CLIENT
     { "update",         hl_cmd_update },
     { "tools",          hl_cmd_tools },
+    { "platform",       hl_cmd_platform },
 #endif
     { "sign-release",   hl_cmd_sign_release },
     { "verify-release", hl_cmd_verify_release },